+++ src/main/scala/com/airbnb/scheduler/jobs/Jobs.scala
@@ -35,8 +35,8 @@ trait BaseJob {
   def lastError: String = ""
   def async: Boolean = false
   def cpus: Double = 0
-  def disk: Int = 0
-  def mem: Int = 0
+  def disk: Double = 0
+  def mem: Double = 0
   def disabled: Boolean = false
   def errorsSinceLastSuccess: Long = 0L
   def uris: Seq[String] = List()
@@ -64,8 +64,8 @@ case class ScheduleBasedJob(
     @JsonProperty override val lastError: String = "",
     @JsonProperty override val async: Boolean = false,
     @JsonProperty override val cpus: Double = 0,
-    @JsonProperty override val disk: Int = 0,
-    @JsonProperty override val mem: Int = 0,
+    @JsonProperty override val disk: Double = 0,
+    @JsonProperty override val mem: Double = 0,
     @JsonProperty override val disabled: Boolean = false,
     @JsonProperty override val errorsSinceLastSuccess: Long = 0L,
     @JsonProperty override val uris: Seq[String] = List(),
@@ -95,8 +95,8 @@ case class DependencyBasedJob(
     @JsonProperty override val lastError: String = "",
     @JsonProperty override val async: Boolean = false,
     @JsonProperty override val cpus: Double = 0,
-    @JsonProperty override val disk: Int = 0,
-    @JsonProperty override val mem: Int = 0,
+    @JsonProperty override val disk: Double = 0,
+    @JsonProperty override val mem: Double = 0,
     @JsonProperty override val disabled: Boolean = false,
     @JsonProperty override val errorsSinceLastSuccess: Long = 0L,
     @JsonProperty override val uris: Seq[String] = List(),
